refactor:fetch로직 추상화한 커스텀 훅 생성(#81)
